gpu - move kernel strings to kernel file
