+++ backends/cuda/ceed-cuda-basis.c
@@ -20,8 +20,8 @@
 #include <cuda_runtime.h>
 #include "ceed-cuda.h"
 #include "ceed-cuda-jit.h"
-#include "kernel-strings/cuda-tensor-basis.h"
-#include "kernel-strings/cuda-non-tensor-basis.h"
+#include "kernels/cuda-tensor-basis.h"
+#include "kernels/cuda-non-tensor-basis.h"
 
 //------------------------------------------------------------------------------
 // Basis apply - tensor
