+++ backends/cuda/ceed-cuda-qfunction-load.cpp
@@ -21,7 +21,7 @@
 #include <string>
 #include "ceed-cuda.h"
 #include "ceed-cuda-jit.h"
-#include "kernel-strings/cuda-qfunction.h"
+#include "kernels/cuda-qfunction.h"
 
 //------------------------------------------------------------------------------
 // Build QFunction kernel
