+++ backends/hip/ceed-hip-restriction.c
@@ -20,7 +20,7 @@
 #include <stddef.h>
 #include "ceed-hip.h"
 #include "ceed-hip-jit.h"
-#include "kernel-strings/hip-restriction.h"
+#include "kernels/hip-restriction.h"
 
 //------------------------------------------------------------------------------
 // Apply restriction
